fix(website): restructure ask question menu
