+++ packages/paste-website/src/pages/components/anchor/index.mdx
@@ -54,14 +54,19 @@ export const pageQuery = graphql`
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/components-anchor--default"
   abstractUrl="https://share.goabstract.com/b202a1d0-0129-4a6e-bc2c-701ae5250955"
   data={props.data.allPasteComponent.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
