+++ packages/paste-website/src/pages/components/button/index.mdx
@@ -47,14 +47,19 @@ export const pageQuery = graphql`
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/components-button--text-only"
   abstractUrl="https://share.goabstract.com/58f1e53e-9f3e-4a04-b84e-44918d1ccdaf"
   data={props.data.allPasteComponent.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
