+++ packages/paste-website/src/pages/components/card/index.mdx
@@ -84,14 +84,19 @@ export const pageQuery = graphql`
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/components-card--default"
   abstractUrl="https://share.goabstract.com/6a1699c0-a54a-47fc-9a4e-84d23b4e42fc?"
   data={props.data.allPasteComponent.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
