+++ packages/paste-website/src/pages/components/heading/index.mdx
@@ -47,14 +47,19 @@ export const pageQuery = graphql`
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/components-heading--all-variants"
   abstractUrl="https://share.goabstract.com/85c81258-90b9-49c8-b34b-9418da5f2f00?"
   data={props.data.allPasteComponent.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
