+++ packages/paste-website/src/pages/components/spinner/index.mdx
@@ -6,6 +6,7 @@ slug: /components/spinner/
 ---
 
 import {graphql} from 'gatsby';
+import {Box} from '@twilio-paste/box';
 import {Stack} from '@twilio-paste/stack';
 import {Spinner} from '@twilio-paste/spinner';
 import Changelog from '@twilio-paste/spinner/CHANGELOG.md';
@@ -45,14 +46,19 @@ export const pageQuery = graphql`
   githubUrl="https://github.com/twilio-labs/paste/tree/master/packages/paste-core/components/spinner"
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/components-spinner--default"
   data={props.data.allPasteComponent.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
