+++ packages/paste-website/src/pages/content/writing-for-components/index.mdx
@@ -5,10 +5,16 @@ slug: /content/writing-for-components/
 ---
 
 import {graphql} from 'gatsby';
+import {Box} from '@twilio-paste/box';
 
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/content/writing-for-components/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -29,7 +35,13 @@ export const pageQuery = graphql`
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
