+++ packages/paste-website/src/pages/icon-system/how-to-add-an-icon/index.mdx
@@ -18,6 +18,11 @@ import {Grid} from '../../../components/grid';
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/icon-system/how-to-add-an-icon/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -27,7 +32,15 @@ export const pageQuery = graphql`
 `;
 
 <contentwrapper>
+
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
+
 <content>
 <h1>{props.pageContext.frontmatter.title}</h1>
