+++ packages/paste-website/src/pages/icon-system/index.mdx
@@ -20,6 +20,11 @@ import Changelog from '@twilio-paste/icons/CHANGELOG.md';
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/icon-system/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -30,7 +35,13 @@ export const pageQuery = graphql`
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
