+++ packages/paste-website/src/pages/libraries/index.mdx
@@ -5,10 +5,16 @@ slug: /libraries/
 ---
 
 import {graphql} from 'gatsby';
+import {Box} from '@twilio-paste/box';
 
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/libraries/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -19,7 +25,13 @@ export const pageQuery = graphql`
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
