+++ packages/paste-website/src/pages/primitives/modal-dialog-primitive/index.mdx
@@ -46,13 +46,20 @@ export const pageQuery = graphql`
   githubUrl="https://github.com/twilio-labs/paste/tree/master/packages/paste-core/primitives/modal-dialog"
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/primitives-modaldialog--custom-overlay-and-content"
   data={props.data.allPastePrimitive.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
+
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
+
 <content>
 
 ## Guidelines
