+++ packages/paste-website/src/pages/primitives/non-modal-dialog-primitive/index.mdx
@@ -57,14 +57,19 @@ export const pageQuery = graphql`
   githubUrl="https://github.com/twilio-labs/paste/tree/master/packages/paste-core/primitives/non-modal-dialog"
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/primitives-non-modal-dialog--simple-non-modal-dialog"
   data={props.data.allPastePrimitive.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
