+++ packages/paste-website/src/pages/primitives/tooltip-primitive/index.mdx
@@ -52,14 +52,19 @@ export const pageQuery = graphql`
   githubUrl="https://github.com/twilio-labs/paste/tree/master/packages/paste-core/primitives/tooltip"
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/primitives-tooltip--simple-tooltip"
   data={props.data.allPastePrimitive.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
