+++ packages/paste-website/src/pages/principles/index.mdx
@@ -5,6 +5,7 @@ slug: /principles/
 ---
 
 import {graphql} from 'gatsby';
+import {Box} from '@twilio-paste/box';
 import {Callout, CalloutTitle, CalloutText} from '../../components/callout';
 import {Grid} from '../../components/grid';
 import {Link} from 'gatsby';
@@ -12,6 +13,11 @@ import {Link} from 'gatsby';
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/principles/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -22,7 +28,13 @@ export const pageQuery = graphql`
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
