+++ packages/paste-website/src/pages/tokens/how-to-compose-custom-ui-with-tokens/index.mdx
@@ -12,6 +12,11 @@ import {Callout, CalloutTitle, CalloutText} from '../../../components/callout';
 export const pageQuery = graphql`
   {
     mdx(fields: {slug: {eq: "/tokens/how-to-compose-custom-ui-with-tokens/"}}) {
+      fileAbsolutePath
+      frontmatter {
+        slug
+        title
+      }
       headings {
         depth
         value
@@ -22,7 +27,13 @@ export const pageQuery = graphql`
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
