+++ packages/paste-website/src/pages/tokens/theme-package/index.mdx
@@ -43,14 +43,19 @@ export const pageQuery = graphql`
   githubUrl="https://github.com/twilio-labs/paste/tree/master/packages/paste-core/components/button"
   storybookUrl="https://twilio-labs.github.io/paste/?path=/story/overview-theme--provider"
   data={props.data.allPasteThemePackage.edges}
-  mdxData={props.data.mdx}
 />
 
 ---
 
 <contentwrapper>
 
+<Box order={2} marginTop="space40" marginLeft="space140" minWidth="size20">
+
+<QuestionMenu data={props.data.mdx} />
+
 <TableOfContents headings={props.data.mdx.headings} />
+
+</Box>
 
 <content>
